requirements.txt
discord.py==2.6.0
aiofiles==0.8.0
aiosqlite==0.17.0
python-dotenv==1.0.1
